[BACKLOG-24638] Make main /api/repo/files endpoints independent of scheduler API
