+++ extensions/src/main/java/org/pentaho/platform/web/http/api/resources/FileResource.java
@@ -45,6 +45,11 @@ import org.pentaho.platform.api.repository2.unified.IRepositoryVersionManager;
 import org.pentaho.platform.api.repository2.unified.IUnifiedRepository;
 import org.pentaho.platform.api.repository2.unified.RepositoryFile;
 import org.pentaho.platform.api.repository2.unified.UnifiedRepositoryAccessDeniedException;
+import org.pentaho.platform.api.repository2.unified.webservices.LocaleMapDto;
+import org.pentaho.platform.api.repository2.unified.webservices.RepositoryFileAclDto;
+import org.pentaho.platform.api.repository2.unified.webservices.RepositoryFileDto;
+import org.pentaho.platform.api.repository2.unified.webservices.RepositoryFileTreeDto;
+import org.pentaho.platform.api.repository2.unified.webservices.StringKeyStringValueDto;
 import org.pentaho.platform.engine.core.output.SimpleOutputHandler;
 import org.pentaho.platform.engine.core.solution.SimpleParameterProvider;
 import org.pentaho.platform.engine.core.system.PentahoSessionHolder;
@@ -58,11 +63,6 @@ import org.pentaho.platform.repository2.ClientRepositoryPaths;
 import org.pentaho.platform.repository2.unified.jcr.JcrRepositoryFileUtils;
 import org.pentaho.platform.repository2.unified.webservices.DefaultUnifiedRepositoryWebService;
 import org.pentaho.platform.repository2.unified.webservices.FileVersioningConfiguration;
-import org.pentaho.platform.api.repository2.unified.webservices.LocaleMapDto;
-import org.pentaho.platform.api.repository2.unified.webservices.RepositoryFileAclDto;
-import org.pentaho.platform.api.repository2.unified.webservices.RepositoryFileDto;
-import org.pentaho.platform.api.repository2.unified.webservices.RepositoryFileTreeDto;
-import org.pentaho.platform.api.repository2.unified.webservices.StringKeyStringValueDto;
 import org.pentaho.platform.security.policy.rolebased.actions.PublishAction;
 import org.pentaho.platform.util.xml.XMLParserFactoryProducer;
 import org.pentaho.platform.web.http.api.resources.services.FileService;
@@ -119,8 +119,6 @@ public class FileResource extends AbstractJaxRSResource {
 
   protected FileService fileService;
 
-  protected SchedulerResource schedulerResource;
-
   protected RepositoryDownloadWhitelist whitelist;
 
   protected static IUnifiedRepository repository;
@@ -136,7 +134,6 @@ public class FileResource extends AbstractJaxRSResource {
 
   public FileResource() {
     fileService = new FileService();
-    schedulerResource = new SchedulerResource();
   }
 
   public FileResource( HttpServletResponse httpServletResponse ) {
@@ -1993,72 +1990,6 @@ public class FileResource extends AbstractJaxRSResource {
     } catch ( Throwable t ) {
       return buildServerErrorResponse( t.getMessage() );
     }
-  }
-
-  /**
-   * Retrieve the list of execute content by lineage id.
-   *
-   * <p><b>Example Request:</b><br />
-   *    GET pentaho/api/repo/files/generatedContentForSchedule?lineageId=
-   * </p>
-   *
-   * @param lineageId the path for the file.
-   *
-   * @return list of RepositoryFileDto objects.
-   *
-   * <p><b>Example Response:</b></p>
-   * <pre function="syntax.xml">
-   * &lt;List&gt;
-   * &lt;repositoryFileDto&gt;
-   * &lt;createdDate&gt;1402911997019&lt;/createdDate&gt;
-   * &lt;fileSize&gt;3461&lt;/fileSize&gt;
-   * &lt;folder&gt;false&lt;/folder&gt;
-   * &lt;hidden&gt;false&lt;/hidden&gt;
-   * &lt;id&gt;ff11ac89-7eda-4c03-aab1-e27f9048fd38&lt;/id&gt;
-   * &lt;lastModifiedDate&gt;1406647160536&lt;/lastModifiedDate&gt;
-   * &lt;locale&gt;en&lt;/locale&gt;
-   * &lt;localePropertiesMapEntries&gt;
-   * &lt;localeMapDto&gt;
-   * &lt;locale&gt;default&lt;/locale&gt;
-   * &lt;properties&gt;
-   * &lt;stringKeyStringValueDto&gt;
-   * &lt;key&gt;file.title&lt;/key&gt;
-   * &lt;value&gt;myFile&lt;/value&gt;
-   * &lt;/stringKeyStringValueDto&gt;
-   * &lt;stringKeyStringValueDto&gt;
-   * &lt;key&gt;jcr:primaryType&lt;/key&gt;
-   * &lt;value&gt;nt:unstructured&lt;/value&gt;
-   * &lt;/stringKeyStringValueDto&gt;
-   * &lt;stringKeyStringValueDto&gt;
-   * &lt;key&gt;title&lt;/key&gt;
-   * &lt;value&gt;myFile&lt;/value&gt;
-   * &lt;/stringKeyStringValueDto&gt;
-   * &lt;stringKeyStringValueDto&gt;
-   * &lt;key&gt;file.description&lt;/key&gt;
-   * &lt;value&gt;myFile Description&lt;/value&gt;
-   * &lt;/stringKeyStringValueDto&gt;
-   * &lt;/properties&gt;
-   * &lt;/localeMapDto&gt;
-   * &lt;/localePropertiesMapEntries&gt;
-   * &lt;locked&gt;false&lt;/locked&gt;
-   * &lt;name&gt;myFile.prpt&lt;/name&gt;&lt;/name&gt;
-   * &lt;originalParentFolderPath&gt;/public/admin&lt;/originalParentFolderPath&gt;
-   * &lt;ownerType&gt;-1&lt;/ownerType&gt;
-   * &lt;path&gt;/public/admin/ff11ac89-7eda-4c03-aab1-e27f9048fd38&lt;/path&gt;
-   * &lt;title&gt;myFile&lt;/title&gt;
-   * &lt;versionId&gt;1.9&lt;/versionId&gt;
-   * &lt;versioned&gt;true&lt;/versioned&gt;
-   * &lt;/repositoryFileAclDto&gt;
-   * &lt;/List&gt;
-   * </pre>
-   */
-  @GET
-  @Path( "/generatedContentForSchedule" )
-  @Produces( { MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON } )
-  @StatusCodes( {
-    @ResponseCode( code = 200, condition = "Successfully got the generated content for schedule" ) } )
-  public List<RepositoryFileDto> doGetGeneratedContentForSchedule( @QueryParam( "lineageId" ) String lineageId ) {
-    return schedulerResource.doGetGeneratedContentForSchedule( lineageId );
   }
 
   /**
